Added the initial code for a Prerender.IO Attribute option instead of using it as a Module.
